+++ src/main/java/micycle/pgs/color/RGB.java
@@ -9,9 +9,24 @@ package micycle.pgs.color;
  */
 public class RGB {
 
+	/** RGB (0, 0, 0) */
 	public static final int BLACK = composeColor(0, 0, 0);
+	/** RGB (255, 255, 255) */
 	public static final int WHITE = composeColor(255, 255, 255);
+	/** RGB (237, 50, 162) */
 	public static final int PINK = composeColor(237, 50, 162);
+	/** RGB (255, 0, 0) */
+	public static final int RED = composeColor(255, 0, 0);
+	/** RGB (0, 255, 0) */
+	public static final int GREEN = composeColor(0, 255, 0);
+	/** RGB (0, 0, 255) */
+	public static final int BLUE = composeColor(0, 0, 255);
+	/** RGB (255, 255, 0) */
+	public static final int YELLOW = composeColor(255, 255, 0);
+	/** RGB (0, 255, 255) */
+	public static final int CYAN = composeColor(0, 255, 255);
+	/** RGB (255, 0, 255) */
+	public static final int FUCHSIA = composeColor(255, 0, 255);
 
 	private static final float INV_255 = 1f / 255f; // used to normalise RGB values to 0...1
 
@@ -19,21 +34,25 @@ public class RGB {
 	}
 
 	/**
+	 * Composes an integer value that represents a color in RGB format.
+	 * 
 	 * @param red   ∈[0, 255]
 	 * @param green ∈[0, 255]
 	 * @param blue  ∈[0, 255]
 	 * @param alpha ∈[0, 255] (where 0 is transparent; 255 is opaque)
-	 * @return
+	 * @return the integer representation of the color in RGB format
 	 */
 	public static int composeColor(final int red, final int green, final int blue, final int alpha) {
 		return alpha << 24 | red << 16 | green << 8 | blue;
 	}
 
 	/**
+	 * Composes an integer value that represents a color in RGB format.
+	 * 
 	 * @param red   ∈[0, 255]
 	 * @param green ∈[0, 255]
 	 * @param blue  ∈[0, 255]
-	 * @return
+	 * @return the integer representation of the color in RGB format
 	 */
 	public static int composeColor(final int red, final int green, final int blue) {
 		return -16777216 | red << 16 | green << 8 | blue;
@@ -93,7 +112,7 @@ public class RGB {
 		return new float[] { (clr >> 16 & 0xff) * INV_255 * alpha, (clr >> 8 & 0xff) * INV_255 * alpha, (clr & 0xff) * INV_255 * alpha,
 				alpha };
 	}
-	
+
 	/**
 	 * Converts hex color strings to Processing integer colors (RRGGBB).
 	 *